DR_Creacion_Funciones_IPS_Admin(Arreglo de migraciones)

+++ database/migrations/2024_08_20_232827_create_departamentoxips_table.php
@@ -16,7 +16,7 @@ class CreateDepartamentoxipsTable extends Migration
         Schema::create('departamentoxips', function (Blueprint $table) {
             $table->integer('cod_depxips');
             $table->integer('cod_departamento');
-            $table->integer('cod_ips');
+            $table->unsignedBigInteger('cod_ips');
 
             $table->primary(['cod_depxips', 'cod_departamento', 'cod_ips']); //Llave compuesta
